# contracts/openapi/keynetra-v0.1.2.yaml
openapi: 3.0.3
info:
  title: KeyNetra
  version: 0.1.2
paths:
...
  /roles/{role_id}/permissions:
    get:
      tags:
      - management
      summary: List Role Permissions
      operationId: list_role_permissions
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          title: Role Id
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_PermissionOut__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
  /roles/{role_id}/permissions/{permission_id}:
    post:
      tags:
      - management
      summary: Add Permission To Role
      operationId: add_permission_to_role
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          title: Role Id
      - name: permission_id
        in: path
        required: true
        schema:
          type: integer
          title: Permission Id
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_PermissionOut_'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags:
      - management
      summary: Remove Permission From Role
      operationId: remove_permission_from_role
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          title: Role Id
      - name: permission_id
        in: path
        required: true
        schema:
          type: integer
          title: Permission Id
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_dict_str__int__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
  /permissions:
    get:
      tags:
      - management
      summary: List Permissions
      operationId: list_permissions
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - nullable: true
          title: Cursor
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_PermissionOut__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      tags:
      - management
      summary: Create Permission
      operationId: create_permission
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_PermissionOut_'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
  /permissions/{permission_id}:
    put:
      tags:
      - management
      summary: Update Permission
      operationId: update_permission
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: integer
          title: Permission Id
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_PermissionOut_'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags:
      - management
      summary: Delete Permission
      operationId: delete_permission
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: integer
          title: Permission Id
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_dict_str__int__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
  /permissions/{permission_id}/roles:
    get:
      tags:
      - management
      summary: List Permission Roles
      operationId: list_permission_roles
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: integer
          title: Permission Id
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_RoleOut__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  schemas:
...
    PermissionCreate:
      properties:
        action:
          type: string
          title: Action
      additionalProperties: false
      type: object
      required:
      - action
      title: PermissionCreate
...
    SuccessResponse_PermissionOut_:
      properties:
        data:
          $ref: '#/components/schemas/PermissionOut'
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          allOf:
          - $ref: '#/components/schemas/ErrorBody'
          nullable: true
          title: Error
      additionalProperties: false
      type: object
      required:
      - data
      title: SuccessResponse[PermissionOut]
...
    SuccessResponse_list_PermissionOut__:
      properties:
        data:
          items:
            $ref: '#/components/schemas/PermissionOut'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          allOf:
          - $ref: '#/components/schemas/ErrorBody'
          nullable: true
          title: Error
      additionalProperties: false
      type: object
      required:
      - data
      title: SuccessResponse[list[PermissionOut]]
...
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
      - error
      properties:
        data:
          nullable: true
          title: Data
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          $ref: '#/components/schemas/ErrorBody'
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    ApiVersionHeader:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        default: v1
      description: 'Requested API version. Supported values: v1.'
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: false
      schema:
        type: string
      description: Tenant key for strict tenancy and management operations.
  headers:
    ApiVersionHeader:
      description: Resolved API version for the response.
      schema:
        type: string
        default: v1
    RequestIdHeader:
      description: Correlation id for the request.
      schema:
        type: string
  responses:
    Error400:
      description: Bad Request
      content:
        application/json:
          schema: &id001
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: bad_request
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: unauthorized
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error403:
      description: Forbidden
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: forbidden
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error404:
      description: Not Found
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: not_found
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error409:
      description: Conflict
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: conflict
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
...
    Error429:
      description: Too Many Requests
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: too_many_requests
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error500:
      description: Internal Server Error
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: internal_error
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
servers:
- url: /
  description: Relative base URL. Configure the SDK/client host per environment.